fix: Outlines No Corners selection box floating point precision.

+++ src/main/java/fi/dy/masa/malilib/render/RenderUtils.java
@@ -2894,19 +2894,19 @@ public class RenderUtils
         final double dy = cameraPos.y;
         final double dz = cameraPos.z;
 
-        final float dxMin = (float) (-dx - expand);
-        final float dyMin = (float) (-dy - expand);
-        final float dzMin = (float) (-dz - expand);
-        final float dxMax = (float) (-dx + expand);
-        final float dyMax = (float) (-dy + expand);
-        final float dzMax = (float) (-dz + expand);
+		final double dxMin = -dx - expand;
+		final double dyMin = -dy - expand;
+		final double dzMin = -dz - expand;
+		final double dxMax = -dx + expand;
+		final double dyMax = -dy + expand;
+		final double dzMax = -dz + expand;
 
-        final float minX = xMin + dxMin;
-        final float minY = yMin + dyMin;
-        final float minZ = zMin + dzMin;
-        final float maxX = xMax + dxMax;
-        final float maxY = yMax + dyMax;
-        final float maxZ = zMax + dzMax;
+		final float minX = (float) (xMin + dxMin);
+		final float minY = (float) (yMin + dyMin);
+		final float minZ = (float) (zMin + dzMin);
+		final float maxX = (float) (xMax + dxMax);
+		final float maxY = (float) (yMax + dyMax);
+		final float maxZ = (float) (zMax + dzMax);
 
         int start, end;
         RenderContext ctx = new RenderContext(() -> "malilib:renderAreaOutlineNoCorners", MaLiLibPipelines.DEBUG_LINES_MASA_SIMPLE_LEQUAL_DEPTH, 0);
@@ -2918,8 +2918,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(start + dxMin, minY, minZ).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
-            buffer.addVertex(end + dxMax, minY, minZ).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
+            buffer.addVertex((float) (start + dxMin), minY, minZ).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
+            buffer.addVertex((float) (end + dxMax), minY, minZ).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMin && pos1.getY() == yMax && pos1.getZ() == zMin) || (pos2.getX() == xMin && pos2.getY() == yMax && pos2.getZ() == zMin) ? xMin + 1 : xMin;
@@ -2927,8 +2927,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(start + dxMin, maxY + 1, minZ).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
-            buffer.addVertex(end + dxMax, maxY + 1, minZ).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
+            buffer.addVertex((float) (start + dxMin), maxY + 1, minZ).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
+            buffer.addVertex((float) (end + dxMax), maxY + 1, minZ).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMin && pos1.getY() == yMin && pos1.getZ() == zMax) || (pos2.getX() == xMin && pos2.getY() == yMin && pos2.getZ() == zMax) ? xMin + 1 : xMin;
@@ -2936,8 +2936,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(start + dxMin, minY, maxZ + 1).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
-            buffer.addVertex(end + dxMax, minY, maxZ + 1).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
+            buffer.addVertex((float) (start + dxMin), minY, maxZ + 1).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
+            buffer.addVertex((float) (end + dxMax), minY, maxZ + 1).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMin && pos1.getY() == yMax && pos1.getZ() == zMax) || (pos2.getX() == xMin && pos2.getY() == yMax && pos2.getZ() == zMax) ? xMin + 1 : xMin;
@@ -2945,8 +2945,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(start + dxMin, maxY + 1, maxZ + 1).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
-            buffer.addVertex(end + dxMax, maxY + 1, maxZ + 1).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
+            buffer.addVertex((float) (start + dxMin), maxY + 1, maxZ + 1).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
+            buffer.addVertex((float) (end + dxMax), maxY + 1, maxZ + 1).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
         }
 
         // Edges along the Y-axis
@@ -2955,8 +2955,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(minX, start + dyMin, minZ).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
-            buffer.addVertex(minX, end + dyMax, minZ).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
+            buffer.addVertex(minX, (float) (start + dyMin), minZ).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
+            buffer.addVertex(minX, (float) (end + dyMax), minZ).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMax && pos1.getY() == yMin && pos1.getZ() == zMin) || (pos2.getX() == xMax && pos2.getY() == yMin && pos2.getZ() == zMin) ? yMin + 1 : yMin;
@@ -2964,8 +2964,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(maxX + 1, start + dyMin, minZ).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
-            buffer.addVertex(maxX + 1, end + dyMax, minZ).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
+            buffer.addVertex(maxX + 1, (float) (start + dyMin), minZ).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
+            buffer.addVertex(maxX + 1, (float) (end + dyMax), minZ).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMin && pos1.getY() == yMin && pos1.getZ() == zMax) || (pos2.getX() == xMin && pos2.getY() == yMin && pos2.getZ() == zMax) ? yMin + 1 : yMin;
@@ -2973,8 +2973,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(minX, start + dyMin, maxZ + 1).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
-            buffer.addVertex(minX, end + dyMax, maxZ + 1).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
+            buffer.addVertex(minX, (float) (start + dyMin), maxZ + 1).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
+            buffer.addVertex(minX, (float) (end + dyMax), maxZ + 1).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMax && pos1.getY() == yMin && pos1.getZ() == zMax) || (pos2.getX() == xMax && pos2.getY() == yMin && pos2.getZ() == zMax) ? yMin + 1 : yMin;
@@ -2982,8 +2982,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(maxX + 1, start + dyMin, maxZ + 1).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
-            buffer.addVertex(maxX + 1, end + dyMax, maxZ + 1).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
+            buffer.addVertex(maxX + 1, (float) (start + dyMin), maxZ + 1).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
+            buffer.addVertex(maxX + 1, (float) (end + dyMax), maxZ + 1).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
         }
 
         // Edges along the Z-axis
@@ -2992,8 +2992,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(minX, minY, start + dzMin).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
-            buffer.addVertex(minX, minY, end + dzMax).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
+            buffer.addVertex(minX, minY, (float) (start + dzMin)).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
+            buffer.addVertex(minX, minY, (float) (end + dzMax)).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMax && pos1.getY() == yMin && pos1.getZ() == zMin) || (pos2.getX() == xMax && pos2.getY() == yMin && pos2.getZ() == zMin) ? zMin + 1 : zMin;
@@ -3001,8 +3001,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(maxX + 1, minY, start + dzMin).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
-            buffer.addVertex(maxX + 1, minY, end + dzMax).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
+            buffer.addVertex(maxX + 1, minY, (float) (start + dzMin)).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
+            buffer.addVertex(maxX + 1, minY, (float) (end + dzMax)).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMin && pos1.getY() == yMax && pos1.getZ() == zMin) || (pos2.getX() == xMin && pos2.getY() == yMax && pos2.getZ() == zMin) ? zMin + 1 : zMin;
@@ -3010,8 +3010,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(minX, maxY + 1, start + dzMin).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
-            buffer.addVertex(minX, maxY + 1, end + dzMax).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
+            buffer.addVertex(minX, maxY + 1, (float) (start + dzMin)).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
+            buffer.addVertex(minX, maxY + 1, (float) (end + dzMax)).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMax && pos1.getY() == yMax && pos1.getZ() == zMin) || (pos2.getX() == xMax && pos2.getY() == yMax && pos2.getZ() == zMin) ? zMin + 1 : zMin;
@@ -3019,8 +3019,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(maxX + 1, maxY + 1, start + dzMin).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
-            buffer.addVertex(maxX + 1, maxY + 1, end + dzMax).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
+            buffer.addVertex(maxX + 1, maxY + 1, (float) (start + dzMin)).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
+            buffer.addVertex(maxX + 1, maxY + 1, (float) (end + dzMax)).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
         }
 
         try
@@ -3029,7 +3029,6 @@ public class RenderUtils
 
             if (meshData != null)
             {
-//                ctx.lineWidth(lineWidth);
                 ctx.draw(meshData, false, true);
                 meshData.close();
             }
